# fix: add max depth limit and max edges limit