--- a/.gitignore
+++ b/.gitignore
@@ -2,3 +2,4 @@
 node_modules/
 npm-debug.log
 yarn.lock
+package-lock.json