--- a/index.d.ts
+++ b/index.d.ts
@@ -1,8 +1,23 @@
-declare function stringify(value: any, replacer?: (key: string, value: any) => any, space?: string | number): string;
+declare function stringify(
+  value: any,
+  replacer?: (key: string, value: any) => any,
+  space?: string | number,
+  options?: { depthLimit: number | undefined; edgesLimit: number | undefined }
+): string;
 
 declare namespace stringify {
-  export function stable(value: any, replacer?: (key: string, value: any) => any, space?: string | number): string;
-  export function stableStringify(value: any, replacer?: (key: string, value: any) => any, space?: string | number): string;
+  export function stable(
+    value: any,
+    replacer?: (key: string, value: any) => any,
+    space?: string | number,
+    options?: { depthLimit: number | undefined; edgesLimit: number | undefined }
+  ): string;
+  export function stableStringify(
+    value: any,
+    replacer?: (key: string, value: any) => any,
+    space?: string | number,
+    options?: { depthLimit: number | undefined; edgesLimit: number | undefined }
+  ): string;
 }
 
 export default stringify;