--- a/test-stable.js
+++ b/test-stable.js
@@ -7,11 +7,9 @@ const stream = require('stream')
 test('circular reference to root', function (assert) {
   const fixture = { name: 'Tywin Lannister' }
   fixture.circle = fixture
-  const expected = s(
-    { circle: '[Circular]', name: 'Tywin Lannister' }
-  )
+  const expected = s({ circle: '[Circular]', name: 'Tywin Lannister' })
   const actual = fss(fixture)
-  assert.is(actual, expected)
+  assert.equal(actual, expected)
   assert.end()
 })
 
@@ -23,50 +21,54 @@ test('circular getter reference to root', function (assert) {
     }
   }
 
-  const expected = s(
-    { circle: '[Circular]', name: 'Tywin Lannister' }
-  )
+  const expected = s({ circle: '[Circular]', name: 'Tywin Lannister' })
   const actual = fss(fixture)
-  assert.is(actual, expected)
+  assert.equal(actual, expected)
   assert.end()
 })
 
 test('nested circular reference to root', function (assert) {
   const fixture = { name: 'Tywin Lannister' }
   fixture.id = { circle: fixture }
-  const expected = s(
-    { id: { circle: '[Circular]' }, name: 'Tywin Lannister' }
-  )
+  const expected = s({ id: { circle: '[Circular]' }, name: 'Tywin Lannister' })
   const actual = fss(fixture)
-  assert.is(actual, expected)
+  assert.equal(actual, expected)
   assert.end()
 })
 
 test('child circular reference', function (assert) {
-  const fixture = { name: 'Tywin Lannister', child: { name: 'Tyrion Lannister' } }
+  const fixture = {
+    name: 'Tywin Lannister',
+    child: { name: 'Tyrion Lannister' }
+  }
   fixture.child.dinklage = fixture.child
   const expected = s({
     child: {
-      dinklage: '[Circular]', name: 'Tyrion Lannister'
+      dinklage: '[Circular]',
+      name: 'Tyrion Lannister'
     },
     name: 'Tywin Lannister'
   })
   const actual = fss(fixture)
-  assert.is(actual, expected)
+  assert.equal(actual, expected)
   assert.end()
 })
 
 test('nested child circular reference', function (assert) {
-  const fixture = { name: 'Tywin Lannister', child: { name: 'Tyrion Lannister' } }
+  const fixture = {
+    name: 'Tywin Lannister',
+    child: { name: 'Tyrion Lannister' }
+  }
   fixture.child.actor = { dinklage: fixture.child }
   const expected = s({
     child: {
-      actor: { dinklage: '[Circular]' }, name: 'Tyrion Lannister'
+      actor: { dinklage: '[Circular]' },
+      name: 'Tyrion Lannister'
     },
     name: 'Tywin Lannister'
   })
   const actual = fss(fixture)
-  assert.is(actual, expected)
+  assert.equal(actual, expected)
   assert.end()
 })
 
@@ -74,10 +76,11 @@ test('circular objects in an array', function (assert) {
   const fixture = { name: 'Tywin Lannister' }
   fixture.hand = [fixture, fixture]
   const expected = s({
-    hand: ['[Circular]', '[Circular]'], name: 'Tywin Lannister'
+    hand: ['[Circular]', '[Circular]'],
+    name: 'Tywin Lannister'
   })
   const actual = fss(fixture)
-  assert.is(actual, expected)
+  assert.equal(actual, expected)
   assert.end()
 })
 
@@ -97,7 +100,7 @@ test('nested circular references in an array', function (assert) {
     ]
   })
   const actual = fss(fixture)
-  assert.is(actual, expected)
+  assert.equal(actual, expected)
   assert.end()
 })
 
@@ -106,7 +109,7 @@ test('circular arrays', function (assert) {
   fixture.push(fixture, fixture)
   const expected = s(['[Circular]', '[Circular]'])
   const actual = fss(fixture)
-  assert.is(actual, expected)
+  assert.equal(actual, expected)
   assert.end()
 })
 
@@ -121,7 +124,7 @@ test('nested circular arrays', function (assert) {
     { bastards: '[Circular]', name: 'Ramsay Bolton' }
   ])
   const actual = fss(fixture)
-  assert.is(actual, expected)
+  assert.equal(actual, expected)
   assert.end()
 })
 
@@ -133,7 +136,7 @@ test('repeated non-circular references in objects', function (assert) {
   }
   const expected = s(fixture)
   const actual = fss(fixture)
-  assert.is(actual, expected)
+  assert.equal(actual, expected)
   assert.end()
 })
 
@@ -142,7 +145,7 @@ test('repeated non-circular references in arrays', function (assert) {
   const fixture = [daenerys, daenerys]
   const expected = s(fixture)
   const actual = fss(fixture)
-  assert.is(actual, expected)
+  assert.equal(actual, expected)
   assert.end()
 })
 
@@ -156,24 +159,28 @@ test('double child circular reference', function (assert) {
   const cloned = clone(fixture)
   const expected = s({
     childA: {
-      dinklage: '[Circular]', name: 'Tyrion Lannister'
+      dinklage: '[Circular]',
+      name: 'Tyrion Lannister'
     },
     childB: {
-      dinklage: '[Circular]', name: 'Tyrion Lannister'
+      dinklage: '[Circular]',
+      name: 'Tyrion Lannister'
     },
     name: 'Tywin Lannister'
   })
   const actual = fss(fixture)
-  assert.is(actual, expected)
+  assert.equal(actual, expected)
 
   // check if the fixture has not been modified
-  assert.deepEqual(fixture, cloned)
+  assert.same(fixture, cloned)
   assert.end()
 })
 
 test('child circular reference with toJSON', function (assert) {
-  // Create a test object that has an overriden `toJSON` property
-  TestObject.prototype.toJSON = function () { return { special: 'case' } }
+  // Create a test object that has an overridden `toJSON` property
+  TestObject.prototype.toJSON = function () {
+    return { special: 'case' }
+  }
   function TestObject (content) {}
 
   // Creating a simple circular object structure
@@ -187,16 +194,22 @@ test('child circular reference with toJSON', function (assert) {
   otherParentObject.otherChildObject.otherParentObject = otherParentObject
 
   // Making sure our original tests work
-  assert.deepEqual(parentObject.childObject.parentObject, parentObject)
-  assert.deepEqual(otherParentObject.otherChildObject.otherParentObject, otherParentObject)
+  assert.same(parentObject.childObject.parentObject, parentObject)
+  assert.same(
+    otherParentObject.otherChildObject.otherParentObject,
+    otherParentObject
+  )
 
   // Should both be idempotent
   assert.equal(fss(parentObject), '{"childObject":{"special":"case"}}')
   assert.equal(fss(otherParentObject), '{"special":"case"}')
 
   // Therefore the following assertion should be `true`
-  assert.deepEqual(parentObject.childObject.parentObject, parentObject)
-  assert.deepEqual(otherParentObject.otherChildObject.otherParentObject, otherParentObject)
+  assert.same(parentObject.childObject.parentObject, parentObject)
+  assert.same(
+    otherParentObject.otherChildObject.otherParentObject,
+    otherParentObject
+  )
 
   assert.end()
 })
@@ -204,14 +217,14 @@ test('child circular reference with toJSON', function (assert) {
 test('null object', function (assert) {
   const expected = s(null)
   const actual = fss(null)
-  assert.is(actual, expected)
+  assert.equal(actual, expected)
   assert.end()
 })
 
 test('null property', function (assert) {
   const expected = s({ f: null })
   const actual = fss({ f: null })
-  assert.is(actual, expected)
+  assert.equal(actual, expected)
   assert.end()
 })
 
@@ -258,7 +271,7 @@ test('nested child circular reference in toJSON', function (assert) {
     }
   })
   const actual = fss(o)
-  assert.is(actual, expected)
+  assert.equal(actual, expected)
   assert.end()
 })
 
@@ -271,7 +284,7 @@ test('circular getters are restored when stringified', function (assert) {
   }
   fss(fixture)
 
-  assert.is(fixture.circle, fixture)
+  assert.equal(fixture.circle, fixture)
   assert.end()
 })
 
@@ -279,13 +292,15 @@ test('non-configurable circular getters use a replacer instead of markers', func
   const fixture = { name: 'Tywin Lannister' }
   Object.defineProperty(fixture, 'circle', {
     configurable: false,
-    get: function () { return fixture },
+    get: function () {
+      return fixture
+    },
     enumerable: true
   })
 
   fss(fixture)
 
-  assert.is(fixture.circle, fixture)
+  assert.equal(fixture.circle, fixture)
   assert.end()
 })
 
@@ -294,20 +309,25 @@ test('getter child circular reference', function (assert) {
     name: 'Tywin Lannister',
     child: {
       name: 'Tyrion Lannister',
-      get dinklage () { return fixture.child }
+      get dinklage () {
+        return fixture.child
+      }
     },
-    get self () { return fixture }
+    get self () {
+      return fixture
+    }
   }
 
   const expected = s({
     child: {
-      dinklage: '[Circular]', name: 'Tyrion Lannister'
+      dinklage: '[Circular]',
+      name: 'Tyrion Lannister'
     },
     name: 'Tywin Lannister',
     self: '[Circular]'
   })
   const actual = fss(fixture)
-  assert.is(actual, expected)
+  assert.equal(actual, expected)
   assert.end()
 })
 
@@ -318,6 +338,67 @@ test('Proxy throwing', function (assert) {
   s.write('', () => {
     assert.end()
   })
-  const actual = fss({ s, p: new Proxy({}, { get() { throw new Error('kaboom') } }) })
-  assert.is(actual, '"[unable to serialize, circular reference is too complex to analyze]"')
+  const actual = fss({ s, p: new Proxy({}, { get () { throw new Error('kaboom') } }) })
+  assert.equal(actual, '"[unable to serialize, circular reference is too complex to analyze]"')
+})
+
+test('depthLimit option - will replace deep objects', function (assert) {
+  const fixture = {
+    name: 'Tywin Lannister',
+    child: {
+      name: 'Tyrion Lannister'
+    },
+    get self () {
+      return fixture
+    }
+  }
+
+  const expected = s({
+    child: '[...]',
+    name: 'Tywin Lannister',
+    self: '[Circular]'
+  })
+  const actual = fss(fixture, undefined, undefined, {
+    depthLimit: 1,
+    edgesLimit: 1
+  })
+  assert.equal(actual, expected)
+  assert.end()
+})
+
+test('edgesLimit option - will replace deep objects', function (assert) {
+  const fixture = {
+    object: {
+      1: { test: 'test' },
+      2: { test: 'test' },
+      3: { test: 'test' },
+      4: { test: 'test' }
+    },
+    array: [
+      { test: 'test' },
+      { test: 'test' },
+      { test: 'test' },
+      { test: 'test' }
+    ],
+    get self () {
+      return fixture
+    }
+  }
+
+  const expected = s({
+    array: [{ test: 'test' }, { test: 'test' }, { test: 'test' }, '[...]'],
+    object: {
+      1: { test: 'test' },
+      2: { test: 'test' },
+      3: { test: 'test' },
+      4: '[...]'
+    },
+    self: '[Circular]'
+  })
+  const actual = fss(fixture, undefined, undefined, {
+    depthLimit: 3,
+    edgesLimit: 3
+  })
+  assert.equal(actual, expected)
+  assert.end()
 })