This supersedes https://github.com/davidmarkclements/fast-safe-stringify/pull/51 by updating the branch to integrate the latest master changes and undoing some unnecessary changes.

See original PR for implementation details