@mcollina 2021-09-07
Is there any performance regression in landing this?